Consolidate metrics/database.ts into one sync file
