+++ glean/src/core/metrics/database.ts
@@ -2,39 +2,116 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import type SynchronousStore from "../../storage/sync.js";
-import type { MetricType, Metrics } from "../index.js";
-import type { Metric } from "../metric.js";
+import type Store from "../storage.js";
+import type { MetricType, Metrics } from "./index.js";
+import type { Metric } from "./metric.js";
+import type { JSONObject, JSONValue } from "../utils.js";
 
-import type { JSONObject, JSONValue } from "../../utils.js";
-import { validateMetricInternalRepresentation } from "../utils.js";
-import { isObject, isUndefined } from "../../utils.js";
-import { Lifetime } from "../lifetime.js";
-import log, { LoggingLevel } from "../../log.js";
-import { Context } from "../../context.js";
-import {
-  createMetricsPayload,
-  METRICS_DATABASE_LOG_TAG,
-  RESERVED_METRIC_IDENTIFIER_PREFIX
-} from "./shared.js";
+import { createMetric, validateMetricInternalRepresentation } from "./utils.js";
+import { isObject, isUndefined } from "../utils.js";
+import { Lifetime } from "./lifetime.js";
+import log, { LoggingLevel } from "../log.js";
+import { Context } from "../context.js";
 
-// See `IMetricsDatabase` for method documentation.
-class MetricsDatabaseSync {
-  private userStore: SynchronousStore;
-  private pingStore: SynchronousStore;
-  private appStore: SynchronousStore;
+/// CONSTANTS ///
+const METRICS_DATABASE_LOG_TAG = "core.Metrics.Database";
+
+// Metrics whose names start with this prefix will
+// not be added to the ping payload.
+//
+// glean_parser rejects metrics with `#` in the name,
+// so this is guaranteed not to clash with user defined metrics.
+const RESERVED_METRIC_NAME_PREFIX = "reserved#";
+
+// The full identifier of internal metrics.
+export const RESERVED_METRIC_IDENTIFIER_PREFIX = `glean.${RESERVED_METRIC_NAME_PREFIX}`;
+
+/// HELPERS ///
+/**
+ * Generates a name for a reserved metric.
+ *
+ * Reserved metrics are not sent in ping payloads.
+ *
+ * @param name The name of the metric.
+ * @returns The name of metrics with proper identification to make it a reserved metric.
+ */
+export function generateReservedMetricIdentifiers(name: string): {
+  category: string;
+  name: string;
+} {
+  return {
+    category: "glean",
+    name: `${RESERVED_METRIC_NAME_PREFIX}${name}`
+  };
+}
+
+/**
+ * Creates the metrics payload from a metrics object with metrics in their internal representation.
+ *
+ * @param v The Metrics object to transform.
+ * @returns A metrics object with metrics in their payload format.
+ */
+export function createMetricsPayload(v: Metrics): Metrics {
+  const result: Metrics = {};
+
+  for (const metricType in v) {
+    const metrics = v[metricType];
+    result[metricType] = {};
+    for (const metricIdentifier in metrics) {
+      const metric = createMetric(metricType, metrics[metricIdentifier]);
+      result[metricType][metricIdentifier] = metric.payload();
+    }
+  }
+
+  return result;
+}
+
+/**
+ * The metrics database is an abstraction layer on top of the underlying storage.
+ *
+ * Metric data is saved to the database in the following format:
+ *
+ * {
+ *  "pingName": {
+ *    "metricType (i.e. boolean)": {
+ *      "metricIdentifier": metricPayload
+ *    }
+ *  }
+ * }
+ *
+ * We have one store in this format for each lifetime: user, ping and application.
+ *
+ */
+class MetricsDatabase {
+  private userStore: Store;
+  private pingStore: Store;
+  private appStore: Store;
 
   constructor() {
-    this.userStore = new Context.platform.Storage("userLifetimeMetrics") as SynchronousStore;
-    this.pingStore = new Context.platform.Storage("pingLifetimeMetrics") as SynchronousStore;
-    this.appStore = new Context.platform.Storage("appLifetimeMetrics") as SynchronousStore;
+    this.userStore = new Context.platform.Storage("userLifetimeMetrics");
+    this.pingStore = new Context.platform.Storage("pingLifetimeMetrics");
+    this.appStore = new Context.platform.Storage("appLifetimeMetrics");
   }
 
   /// PUBLIC ///
+  /**
+   * Records a given value to a given metric.
+   * Will overwrite in case there is already a value in there.
+   *
+   * @param metric The metric to record to.
+   * @param value The value we want to record to the given metric.
+   */
   record(metric: MetricType, value: Metric<JSONValue, JSONValue>): void {
     this.transform(metric, () => value);
   }
 
+  /**
+   * Records a given value to a given metric,
+   * by applying a transformation function on the value currently persisted.
+   *
+   * @param metric The metric to record to.
+   * @param transformFn The transformation function to apply to the currently persisted value.
+   */
   transform(
     metric: MetricType,
     transformFn: (v?: JSONValue) => Metric<JSONValue, JSONValue>
@@ -45,7 +122,7 @@ class MetricsDatabaseSync {
 
     const store = this.chooseStore(metric.lifetime);
 
-    const storageKey = metric.identifierSync();
+    const storageKey = metric.identifier();
 
     for (const ping of metric.sendInPings) {
       const finalTransformFn = (v?: JSONValue): JSONValue => transformFn(v).get();
@@ -53,6 +130,16 @@ class MetricsDatabaseSync {
     }
   }
 
+  /**
+   * Checks if anything was stored for the provided metric.
+   *
+   * @param lifetime the metric `Lifetime`.
+   * @param ping the ping storage to search in.
+   * @param metricType the type of the metric.
+   * @param metricIdentifier the metric identifier.
+   * @returns `true` if the metric was found (regardless of the validity of the
+   *          stored data), `false` otherwise.
+   */
   hasMetric(
     lifetime: Lifetime,
     ping: string,
@@ -64,6 +151,15 @@ class MetricsDatabaseSync {
     return !isUndefined(value);
   }
 
+  /**
+   * Counts the number of stored metrics with an id starting with a specific identifier.
+   *
+   * @param lifetime the metric `Lifetime`.
+   * @param ping the ping storage to search in.
+   * @param metricType the type of the metric.
+   * @param metricIdentifier the metric identifier.
+   * @returns the number of stored metrics with their id starting with the given identifier.
+   */
   countByBaseIdentifier(
     lifetime: Lifetime,
     ping: string,
@@ -79,9 +175,31 @@ class MetricsDatabaseSync {
     return Object.keys(pingStorage).filter((n) => n.startsWith(metricIdentifier)).length;
   }
 
+  /**
+   * Gets  and validates the persisted payload of a given metric in a given ping.
+   *
+   * If the persisted value is invalid for the metric we are attempting to retrieve,
+   * the persisted value is deleted and `undefined is returned.
+   *
+   * This behaviour is not consistent with what the Glean SDK does, but this is on purpose.
+   * On the Glean SDK we panic when we can't serialize the persisted value,
+   * that is because this is an extremely unlikely situation for that environment.
+   *
+   * Since Glean.js will run on the browser, it is easy for a consumers / developers
+   * to mess with the storage which makes this sort of errors plausible.
+   * That is why we choose to not panic and simply delete the corrupted data here.
+   *
+   * Note: This is not a strong guard against consumers / developers messing with the storage on their own.
+   * Currently Glean.js does not include mechanisms to reliably prevent that.
+   *
+   * @param ping The ping from which we want to retrieve the given metric.
+   * @param metric An object containing the information about the metric to retrieve.
+   * @returns The payload persisted for the given metric,
+   *          `undefined` in case the metric has not been recorded yet or the found values in invalid.
+   */
   getMetric<T extends JSONValue>(ping: string, metric: MetricType): T | undefined {
     const store = this.chooseStore(metric.lifetime);
-    const storageKey = metric.identifierSync();
+    const storageKey = metric.identifier();
     const value = store.get([ping, metric.type, storageKey]);
     if (!isUndefined(value) && !validateMetricInternalRepresentation<T>(metric.type, value)) {
       log(
@@ -96,6 +214,14 @@ class MetricsDatabaseSync {
     }
   }
 
+  /**
+   * Gets all of the persisted metrics related to a given ping.
+   *
+   * @param ping The name of the ping to retrieve.
+   * @param clearPingLifetimeData Whether or not to clear the ping lifetime metrics retrieved.
+   * @returns An object containing all the metrics recorded to the given ping,
+   *          `undefined` in case the ping doesn't contain any recorded metrics.
+   */
   getPingMetrics(ping: string, clearPingLifetimeData: boolean): Metrics | undefined {
     const userData = this.getCorrectedPingData(ping, Lifetime.User);
     const pingData = this.getCorrectedPingData(ping, Lifetime.Ping);
@@ -132,12 +258,21 @@ class MetricsDatabaseSync {
     }
   }
 
+  /**
+   * Clears currently persisted data for a given lifetime.
+   *
+   * @param lifetime The lifetime to clear.
+   * @param ping The ping to clear data from. When omitted, data from all pings will be cleared.
+   */
   clear(lifetime: Lifetime, ping?: string): void {
     const store = this.chooseStore(lifetime);
     const storageIndex = ping ? [ping] : [];
     store.delete(storageIndex);
   }
 
+  /**
+   * Clears all persisted metrics data.
+   */
   clearAll(): void {
     this.userStore.delete([]);
     this.pingStore.delete([]);
@@ -152,7 +287,7 @@ class MetricsDatabaseSync {
    * @returns The store related to the given lifetime.
    * @throws If the provided lifetime does not have a related store.
    */
-  private chooseStore(lifetime: Lifetime): SynchronousStore {
+  private chooseStore(lifetime: Lifetime): Store {
     switch (lifetime) {
     case Lifetime.User:
       return this.userStore;
@@ -264,4 +399,4 @@ class MetricsDatabaseSync {
   }
 }
 
-export default MetricsDatabaseSync;
+export default MetricsDatabase;
